+++ otter/generate/templates/r/setup.sh
@@ -11,27 +11,27 @@ apt-get install -y libnlopt-dev cmake libfreetype6-dev libpng-dev libtiff5-dev l
     build-essential libcurl4-gnutls-dev libxml2-dev libssl-dev libgit2-dev texlive-lang-chinese \
     libxft-dev
 
-# install conda
+# install mamba
 if [ $(uname -p) = "arm" ] || [ $(uname -p) = "aarch64" ] ; \
-    then wget -nv https://repo.anaconda.com/miniconda/Miniconda3-latest-Linux-aarch64.sh \
-        -O {{ autograder_dir }}/source/miniconda_install.sh ; \
-    else wget -nv https://repo.anaconda.com/miniconda/Miniconda3-latest-Linux-x86_64.sh \
-        -O {{ autograder_dir }}/source/miniconda_install.sh ; \
+    then wget -nv https://github.com/conda-forge/miniforge/releases/latest/download/Mambaforge-Linux-aarch64.sh \
+        -O {{ autograder_dir }}/source/mamba_install.sh ; \
+    else wget -nv https://github.com/conda-forge/miniforge/releases/latest/download/Mambaforge-Linux-x86_64.sh \
+        -O {{ autograder_dir }}/source/mamba_install.sh ; \
 fi
-chmod +x {{ autograder_dir }}/source/miniconda_install.sh
-{{ autograder_dir }}/source/miniconda_install.sh -b
-echo "export PATH=/root/miniconda3/bin:\$PATH" >> /root/.bashrc
+chmod +x {{ autograder_dir }}/source/mamba_install.sh
+{{ autograder_dir }}/source/mamba_install.sh -b
+echo "export PATH=/root/mambaforge/bin:\$PATH" >> /root/.bashrc
 
-export PATH=/root/miniconda3/bin:$PATH
+export PATH=/root/mambaforge/bin:$PATH
 export TAR="/bin/tar"
 
-# install dependencies with conda{% if channel_priority_strict %}
-conda config --set channel_priority strict{% endif %}
-conda env create -f {{ autograder_dir }}/source/environment.yml{% if has_r_requirements %}
-conda run -n {{ otter_env_name }} Rscript {{ autograder_dir }}/source/requirements.r{% endif %}
+# install dependencies with mamba{% if channel_priority_strict %}
+mamba config --set channel_priority strict{% endif %}
+mamba env create -f {{ autograder_dir }}/source/environment.yml{% if has_r_requirements %}
+mamba run -n {{ otter_env_name }} Rscript {{ autograder_dir }}/source/requirements.r{% endif %}
 
-# set conda shell
-conda init --all
+# set mamba shell
+mamba init --all
 
-# install ottr; not sure why it needs to happen twice but whatever
-conda run -n otter-env Rscript -e 'install.packages("https://cran.r-project.org/package=ottr&version={{ ottr_version }}", dependencies=TRUE, repos=NULL)'
+# install ottr
+mamba run -n otter-env Rscript -e 'install.packages("https://cran.r-project.org/package=ottr&version={{ ottr_version }}", dependencies=TRUE, repos=NULL)'