+++ test/test_generate/files/autograder-r-correct/setup.sh
@@ -11,25 +11,25 @@ apt-get install -y libnlopt-dev cmake libfreetype6-dev libpng-dev libtiff5-dev l
     build-essential libcurl4-gnutls-dev libxml2-dev libssl-dev libgit2-dev texlive-lang-chinese \
     libxft-dev
 
-# install conda
+# install mamba
 if [ $(uname -p) = "arm" ] || [ $(uname -p) = "aarch64" ] ; \
-    then wget -nv https://repo.anaconda.com/miniconda/Miniconda3-latest-Linux-aarch64.sh \
-        -O /autograder/source/miniconda_install.sh ; \
-    else wget -nv https://repo.anaconda.com/miniconda/Miniconda3-latest-Linux-x86_64.sh \
-        -O /autograder/source/miniconda_install.sh ; \
+    then wget -nv https://github.com/conda-forge/miniforge/releases/latest/download/Mambaforge-Linux-aarch64.sh \
+        -O /autograder/source/mamba_install.sh ; \
+    else wget -nv https://github.com/conda-forge/miniforge/releases/latest/download/Mambaforge-Linux-x86_64.sh \
+        -O /autograder/source/mamba_install.sh ; \
 fi
-chmod +x /autograder/source/miniconda_install.sh
-/autograder/source/miniconda_install.sh -b
-echo "export PATH=/root/miniconda3/bin:\$PATH" >> /root/.bashrc
+chmod +x /autograder/source/mamba_install.sh
+/autograder/source/mamba_install.sh -b
+echo "export PATH=/root/mambaforge/bin:\$PATH" >> /root/.bashrc
 
-export PATH=/root/miniconda3/bin:$PATH
+export PATH=/root/mambaforge/bin:$PATH
 export TAR="/bin/tar"
 
-# install dependencies with conda
-conda env create -f /autograder/source/environment.yml
+# install dependencies with mamba
+mamba env create -f /autograder/source/environment.yml
 
-# set conda shell
-conda init --all
+# set mamba shell
+mamba init --all
 
-# install ottr; not sure why it needs to happen twice but whatever
-conda run -n otter-env Rscript -e 'install.packages("https://cran.r-project.org/package=ottr&version=1.2.0", dependencies=TRUE, repos=NULL)'
+# install ottr
+mamba run -n otter-env Rscript -e 'install.packages("https://cran.r-project.org/package=ottr&version=1.2.0", dependencies=TRUE, repos=NULL)'